backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1